+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - Adds (opt-out) experiment to disable cleaning up containers after a functions deploy (#6861)
+- Fix Next.js image optimization check in app directory for Windows (#6930)
 - Add support to next.config.mjs (#6933)